add Authorisation and relation many to one with user and task entity

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { TaskRepository } from './task.repository';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,8 +15,8 @@ export class TasksService {
   ) {
   }
 
-  async getTaskById(id: number): Promise<Task | undefined> {
-    const exist = await this.taskRepository.findOne(id)
+  async getTaskById(id: number, user: User): Promise<Task | undefined> {
+    const exist = await this.taskRepository.findOne({ where: { id, userId: user.id } })
     if (!exist) {
       throw new NotFoundException(`Task with ID ${id} not found`)
     }
@@ -23,23 +24,26 @@ export class TasksService {
     return exist
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto)
+  async createTask(
+    createTaskDto: CreateTaskDto,
+    user: User,
+    ): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user)
   }
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return await this.taskRepository.getTasks(filterDto)
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return await this.taskRepository.getTasks(filterDto, user)
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const res = await this.taskRepository.delete(id)
+  async deleteTask(id: number, user: User): Promise<void> {
+    const res = await this.taskRepository.delete({ id, userId: user.id })
     if (res.affected === 0) {
       throw new NotFoundException(`Task with ID ${id} not found`)
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id)
+  async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user)
     task.status = status
     await task.save()
     return task
